TargetTrialEmulation.jl Documentation

TargetTrialEmulation.seqtrial — Function

seqtrial(df::DataFrame, covariates::Array{Symbol,1})

Set up treatment arms of sequential target trial. Confounders are fixed to baseline values.

Keyword Arguments

df::DataFrame: DataFrame with columns :id, :period, :eligible, :treatment.

covariates::Array{Symbol,1}: Array of covariates (that are in the DataFrame) to be fixed to baseline values.

Output

trials_dict: Dict with keys as period numbers and values as DataFrames with the following columns:

:id: ID of the patient.

:period: Timepoint of the observation.

:eligible: Indicator if patient is eligible.

:treatment: Indicator if patient is treated.

:trialnr: Trial number.

:fup: Follow-up time.

:baseline_treatment: Indicator if patient is treated at baseline.

covariates: Covariates fixed to baseline values.

Example

using DataFrames
df = DataFrame(id = [1, 1, 2, 2, 3, 3, 4, 4],
               period = [1, 2, 1, 2, 1, 2, 1, 2],
               eligible = [1, 1, 1, 1, 1, 1, 1, 1],
               treatment = [0, 1, 0, 1, 0, 1, 0, 1],
               age = [20, 20, 30, 30, 40, 40, 50, 50])
covariates = [:age]
trials_dict = seqtrial(df, covariates)

TargetTrialEmulation.TTE — Function

TTE(df::DataFrame, outcome::Symbol, treatment::Symbol, period::Symbol, eligible::Symbol, censored::Symbol, covariates::Array{Symbol,1}, save_w_model::Bool = false)

This is a wrapper function to emulates target trials using the sequential trial emulation method.

Keyword Arguments

outcome::Symbol: Outcome column.

treatment::Symbol: Treatment column.

period::Symbol: Period column.

eligible::Symbol: Eligible column.

censored::Symbol: Censored column.

covariates::Array{Symbol,1}: Array of covariates (that are in the DataFrame) to be fixed to baseline values.

save_w_model::Bool: Indicator if the weighting model should be saved.

Output

df: DataFrame with the following columns:

:id: ID of the patient.

:period: Timepoint of the observation.

:eligible: Indicator if patient is eligible.

:treatment: Indicator if patient is treated.

:trialnr: Trial number.

:fup: Follow-up time.

:baseline_treatment: Indicator if patient is treated at baseline.

covariates: Covariates fixed to baseline values.

out_model: Outcome model.

model_num: Numerator model (only if save_w_model is true).

model_denom: Denominator model (only if save_w_model is true).

Example

TargetTrialEmulation.art_censor — Function

art_censor(dict::Dict{Int64, DataFrame})

Artificially censor individuals who deviate from the treatment.

Keyword Arguments

dict::Dict{Int64, DataFrame}: Dictionary with keys as period numbers and values as DataFrames with the following columns:

:id: ID of the patient.

:period: Timepoint of the observation.

:eligible: Indicator if patient is eligible.

:treatment: Indicator if patient is treated.

:trialnr: Trial number.

:fup: Follow-up time.

:baseline_treatment: Indicator if patient is treated at baseline.

covariates: Covariates fixed to baseline values.

Output

dict: Dictionary with keys as period numbers and values as DataFrames with the following columns:

:id: ID of the patient.

:period: Timepoint of the observation.

:eligible: Indicator if patient is eligible.

:treatment: Indicator if patient is treated.

:trialnr: Trial number.

:fup: Follow-up time.

:baseline_treatment: Indicator if patient is treated at baseline.

covariates: Covariates fixed to baseline values.

censor: Indicator if patient is censored.

Example